backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic
PyMuPDF
langchain
werkzeug
psycopg2-binary
chromadb